+++ src/client/java/net/insomniacs/nucleus/api/geo_animation/data/GeoScale.java
@@ -2,17 +2,17 @@ package net.insomniacs.nucleus.api.geo_animation.data;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import net.insomniacs.nucleus.utils.Vec3f;
 import net.minecraft.client.render.entity.animation.AnimationHelper;
 import net.minecraft.client.render.entity.animation.Keyframe;
 import net.minecraft.client.render.entity.animation.Transformation;
+import net.minecraft.util.math.Vec3d;
 
 public record GeoScale(
-        Vec3f scale
+        Vec3d scale
 ) {
 
     public static final Codec<GeoScale> CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            Vec3f.CODEC.fieldOf("scale").forGetter(null)
+            Vec3d.CODEC.fieldOf("scale").forGetter(null)
     ).apply(instance, GeoScale::new));
 
     public Keyframe toKeyframe(float timestamp) {
